editoast: move sprite_config to `editoast::generated_data`
