+++ editoast/src/generated_data/signal.rs
@@ -14,9 +14,9 @@ use super::utils::InvolvedObjects;
 use super::GeneratedData;
 use crate::diesel::ExpressionMethods;
 use crate::error::Result;
+use crate::generated_data::sprite_config::SpriteConfig;
+use crate::generated_data::sprite_config::SpriteConfigs;
 use crate::infra_cache::InfraCache;
-use crate::schema::sprite_config::SpriteConfig;
-use crate::schema::sprite_config::SpriteConfigs;
 use crate::schema::LogicalSignal;
 use crate::schema::ObjectType;
 use crate::tables::infra_layer_signal::dsl;
