+++ editoast/src/generated_data/sprite_config.rs
