+++ editoast/src/schema/mod.rs
@@ -12,7 +12,6 @@ pub mod rolling_stock;
 mod route;
 mod signal;
 mod speed_section;
-pub mod sprite_config;
 mod switch;
 mod switch_type;
 pub mod track_section;
